Mostly fixing sound popping.
Popping was caused by changing the duration in the middle of bipping with a lower value. Such change would create a discontinuity in the sound amplitude and generate popping with the current implementation. A better implementation could be done (it's CHIP-8, shall not be necesary). Ensuring the length is not modified to a lower length fix the problem for most of the cases. I can imagine situation where the poping would remain but pretty sure not so much a problem for CHIP-8. Though it could be improved